requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-JWT-Extended
pytz
bcrypt
python-dotenv
gunicorn
ldap3
sphinx
Flask-Migrate